Remove dist dir before building wheel package
If any type of artifact was created by a preceding step (i.e. egg
building), when publishing the artifact, twine tries to upload that too.
This causes problems if the version of that artifact is not in line with
what Runway enforces. This commit fixes this by forcing the removal of
the dist directory before building the final package. This does mean
that if you want an egg and a package, the egg should be uploaded first,
and only after that should the package be built