Drop Kubernetes image rolling update task
Rolling image updates can be done using the kubectl CLI, but are in fact
a strategy that can/should be included in your deployment.yml. If you do
this, you can simply use the normal patch/create logic for a Kubernetes
deployment. Given this, we can drop the kubernetes_image_rolling_update
task from Takeoff, and defer this logic to the user, to be specified in
his/her Kubernetes yaml.