Fix base64 encoding of k8s templated values
Until now, Takeoff made some assumptions regarding base64 encoding,
which in some cases led to very awkward situations. The best example was
that in order to render a value from Azure Keyvault into the value of a
Kubernetes secret, the value in the Azure Keyvault would need to be
base64 encoded. This commit makes 2 general assumptions:
- template values that are fetched from the keyvault are to be base64
encoded, as they are secret.
- custom values that are provided via Takeoff config are not to be
base64 encoded, as they are already in plain text in deployment.yml

In order to allow maximum flexibility, a custom jinja2 filter is also
added, that allows users to deviate from the assumptions mentioned
above, by enforcing base64 encoding or decoding in the kubernetes
configuration. This option should be used with a lot of care, as it
could result in leaking secrets.