Make multiple k8s tasks possible
Until now, the assumption was that the .kube config directory did not exist, because
takeoff was (usually) being run in a docker container. Although this is true, it made it
impossible to run multiple deploy_to_kubernetes tasks, as the directory was created in
the first task execution, and threw an error in the second. This change only creates the
.kube directory in the container if it does not already exist