Fix failing black linting
Black was failing on master. This fixes that. The reason this failed
on master but not on the original branch was that there was some time
between the creation of the original PR and the actual merging. This
led to a new version of black being released, which threw different
errors. This happened because we had not fixed a version for the black
package in our CI. This has now been frozen.